// packages/frontend-api-grpc/src/proto/user.ts
export interface User {
  /**
   * @generated from protobuf field: string id = 1;
   */
  id: string;
  /**
   * @generated from protobuf field: google.protobuf.Timestamp createdAt = 2;
   */
  createdAt?: Timestamp;
  /**
   * @generated from protobuf field: google.protobuf.Timestamp updatedAt = 3;
   */
  updatedAt?: Timestamp;
  /**
   * @generated from protobuf field: string name = 4;
   */
  name: string;
  /**
   * @generated from protobuf field: whiterabbit.user.Role role = 5;
   */
  role: Role;
  /**
   * @generated from protobuf field: map<string, string> authIds = 6;
   */
  authIds: {
    [key: string]: string;
  };
}
// ...
export interface CreateCommand {
  /**
   * @generated from protobuf field: optional string targetId = 1;
   */
  targetId?: string;
  /**
   * @generated from protobuf field: string name = 2;
   */
  name: string;
  /**
   * @generated from protobuf field: optional whiterabbit.user.Role role = 3;
   */
  role?: Role;
  /**
   * @generated from protobuf field: whiterabbit.user.NullableAuthIds authIds = 4;
   */
  authIds?: NullableAuthIds;
}
/**
 * @generated from protobuf message whiterabbit.user.UpdateCommand
 */
export interface UpdateCommand {
  /**
   * @generated from protobuf field: string targetId = 1;
   */
  targetId: string;
  /**
   * @generated from protobuf field: optional string name = 2;
   */
  name?: string;
  /**
   * @generated from protobuf field: optional whiterabbit.user.Role role = 3;
   */
  role?: Role;
  /**
   * @generated from protobuf field: whiterabbit.user.NullableAuthIds authIds = 4;
   */
  authIds?: NullableAuthIds;
}
/**
 * @generated from protobuf message whiterabbit.user.NullableAuthIds
 */
export interface NullableAuthIds {
  /**
   * @generated from protobuf field: map<string, string> values = 1;
   */
  values: {
    [key: string]: string;
  };
}
/**
 * @generated from protobuf message whiterabbit.user.DeleteCommand
 */
export interface DeleteCommand {
  /**
   * @generated from protobuf field: string targetId = 1;
   */
  targetId: string;
}
/**
 * @generated from protobuf message whiterabbit.user.Command
 */
export interface Command {
  /**
   * @generated from protobuf oneof: command
   */
  command:
    | {
        oneofKind: "create";
        /**
         * @generated from protobuf field: whiterabbit.user.CreateCommand create = 1;
         */
        create: CreateCommand;
      }
    | {
        oneofKind: "update";
        /**
         * @generated from protobuf field: whiterabbit.user.UpdateCommand update = 2;
         */
        update: UpdateCommand;
      }
    | {
        oneofKind: "delete";
        /**
         * @generated from protobuf field: whiterabbit.user.DeleteCommand delete = 3;
         */
        delete: DeleteCommand;
      }
    | {
        oneofKind: undefined;
      };
}
/**
 * @generated from protobuf message whiterabbit.user.Commands
 */
export interface Commands {
  /**
   * @generated from protobuf field: repeated whiterabbit.user.Command commands = 1;
   */
  commands: Command[];
}
/**
 * @generated from protobuf enum whiterabbit.user.Role
 */
export enum Role {
  /**
   * @generated from protobuf enum value: USER = 0;
   */
  USER = 0,
  /**
   * @generated from protobuf enum value: ADMIN = 1;
   */
  ADMIN = 1,
  /**
   * @generated from protobuf enum value: OWNER = 2;
   */
  OWNER = 2,
}
// @generated message type with reflection information, may provide speed optimized methods
class User$Type extends MessageType<User> {
  constructor() {
    super("whiterabbit.user.User", [
      { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
      { no: 2, name: "createdAt", kind: "message", T: () => Timestamp },
      { no: 3, name: "updatedAt", kind: "message", T: () => Timestamp },
      { no: 4, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
      {
        no: 5,
        name: "role",
        kind: "enum",
        T: () => ["whiterabbit.user.Role", Role],
      },
      {
        no: 6,
        name: "authIds",
        kind: "map",
        K: 9 /*ScalarType.STRING*/,
        V: { kind: "scalar", T: 9 /*ScalarType.STRING*/ },
      },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message whiterabbit.user.User
 */
export const User = new User$Type();
// ...
class CreateCommand$Type extends MessageType<CreateCommand> {
  constructor() {
    super("whiterabbit.user.CreateCommand", [
      {
        no: 1,
        name: "targetId",
        kind: "scalar",
        opt: true,
        T: 9 /*ScalarType.STRING*/,
      },
      { no: 2, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
      {
        no: 3,
        name: "role",
        kind: "enum",
        opt: true,
        T: () => ["whiterabbit.user.Role", Role],
      },
      { no: 4, name: "authIds", kind: "message", T: () => NullableAuthIds },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message whiterabbit.user.CreateCommand
 */
export const CreateCommand = new CreateCommand$Type();
// @generated message type with reflection information, may provide speed optimized methods
class UpdateCommand$Type extends MessageType<UpdateCommand> {
  constructor() {
    super("whiterabbit.user.UpdateCommand", [
      { no: 1, name: "targetId", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
      {
        no: 2,
        name: "name",
        kind: "scalar",
        opt: true,
        T: 9 /*ScalarType.STRING*/,
      },
      {
        no: 3,
        name: "role",
        kind: "enum",
        opt: true,
        T: () => ["whiterabbit.user.Role", Role],
      },
      { no: 4, name: "authIds", kind: "message", T: () => NullableAuthIds },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message whiterabbit.user.UpdateCommand
 */
export const UpdateCommand = new UpdateCommand$Type();
// ...
export const NullableAuthIds = new NullableAuthIds$Type();
// @generated message type with reflection information, may provide speed optimized methods
class DeleteCommand$Type extends MessageType<DeleteCommand> {
  constructor() {
    super("whiterabbit.user.DeleteCommand", [
      { no: 1, name: "targetId", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message whiterabbit.user.DeleteCommand
 */
export const DeleteCommand = new DeleteCommand$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Command$Type extends MessageType<Command> {
  constructor() {
    super("whiterabbit.user.Command", [
      {
        no: 1,
        name: "create",
        kind: "message",
        oneof: "command",
        T: () => CreateCommand,
      },
      {
        no: 2,
        name: "update",
        kind: "message",
        oneof: "command",
        T: () => UpdateCommand,
      },
      {
        no: 3,
        name: "delete",
        kind: "message",
        oneof: "command",
        T: () => DeleteCommand,
      },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message whiterabbit.user.Command
 */
export const Command = new Command$Type();